- Explicitly define Point<double> and reformulate methods appropriately - Explicitly instantiate various templated classes (including PlaneMap where   used, not where declared) - Circumvented 'definition of dllimport' errors in various ways - Fixed indentation, whitespace, spelling, etc - Disable 'not all code paths return a value' warning because it appears when   not true a lot - Store string lengths as uint32s, not uint8s in Archives - Don't use method parameters as stack-allocated array dimensions - Added some not-really-necessary return statments in inlined methods to   appease MSVC - Wrapped gettimeofday in a function that works in Windows too - Corrected z1's WinSock code in the light of compiler errors - Use inet_ntoa rather than inet_ntop under Windows - Renamed libsakusen_comms project to libsakusen-comms to match - Added new files to libsakusen-comms project - Renamed Socket::setAsyncronous to more accurate Socket::setNonBlocking - Use seperate import/export macros for each dll, not one for all of them   (this is what fixed the linking errors that Simeon had been confused by) - Added library directory where my version of the STLport .lib file ended up - Removed various files from projects when they had been inappropriately   included - Corrected arithmetic in doLoadTest() in libsakusen-server-test - Moved a non-templated method from region-methods.h to region.cpp - Removed now-unnecessary includes from map.h - Clarified precedence in stringUtils_findPreviousCharStart
git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@293 6f43db7c-3df8-0310-95f7-eae406835aab

+++ libsakusen/iarchive.cpp
@@ -68,7 +68,7 @@ IArchive& IArchive::operator>>(double& d)
 
 IArchive& IArchive::operator>>(String& s)
 {
-  uint8 length;
+  uint32 length;
   *this >> length;
   assertLength(length);
   s = String(
